requirements.txt
# DEEPVISION - Deepfake Detection System
# Requirements for the 6-month major project

# Core Dependencies
streamlit>=1.35.0
streamlit-webrtc>=0.47.0
python>=3.10.0

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.15.0
keras>=3.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0
scikit-image>=0.21.0

# Video Processing
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
altair>=5.1.0

# Machine Learning Utilities
scikit-learn>=1.3.0
tqdm>=4.66.0

# Model Zoo & Pretrained
timm>=0.9.0
transformers>=4.35.0

# UI Enhancements
streamlit-option-menu>=0.3.0
streamlit-card>=0.0.4

# Utilities
python-magic>=0.4.27
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0

# Logging & Monitoring
loguru>=0.7.0

# Web Server (for deployment)
uvicorn>=0.24.0
fastapi>=0.104.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation
sphinx>=7.2.0

# Type Checking
mypy>=1.6.0